+++ pkg-manager/plugin-commands-installation/src/add.ts
@@ -227,12 +227,13 @@ export async function handler (
         })
       }
     }
-    opts.rootProjectManifest = opts.rootProjectManifest ?? {}
-    opts.rootProjectManifest.pnpm = opts.rootProjectManifest.pnpm ?? {}
-    opts.rootProjectManifest.pnpm.onlyBuiltDependencies = Array.from(new Set([
-      ...(opts.rootProjectManifest.pnpm.onlyBuiltDependencies ?? []),
+    opts.onlyBuiltDependencies = Array.from(new Set([
+      ...(opts.onlyBuiltDependencies ?? []),
       ...opts.allowBuild,
     ])).sort((a, b) => a.localeCompare(b))
+    opts.rootProjectManifest = opts.rootProjectManifest ?? {}
+    opts.rootProjectManifest.pnpm = opts.rootProjectManifest.pnpm ?? {}
+    opts.rootProjectManifest.pnpm.onlyBuiltDependencies = opts.onlyBuiltDependencies
     const writeProjectManifest = await createProjectManifestWriter(opts.rootProjectManifestDir)
     await writeProjectManifest(opts.rootProjectManifest)
   }