+++ __utils__/jest-config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pnpm/jest-config
 
+## 1.0.9
+
+### Patch Changes
+
+- @pnpm/worker@1000.0.7
+
 ## 1.0.8
 
 ### Patch Changes