+++ cache/commands/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pnpm/cache.commands
 
+## 1000.0.13
+
+### Patch Changes
+
+- Updated dependencies [b8b0c68]
+- Updated dependencies [d965748]
+  - @pnpm/cache.api@1000.0.12
+  - @pnpm/config@1002.5.0
+  - @pnpm/cli-utils@1000.0.13
+
 ## 1000.0.12
 
 ### Patch Changes