+++ completion/plugin-commands-completion/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm/plugin-commands-completion
 
+## 1000.0.13
+
+### Patch Changes
+
+- @pnpm/workspace.find-packages@1000.0.13
+- @pnpm/cli-utils@1000.0.13
+- @pnpm/workspace.read-manifest@1000.1.1
+
 ## 1000.0.12
 
 ### Patch Changes