+++ env/system-node-version/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pnpm/env.system-node-version
 
+## 1000.0.4
+
+### Patch Changes
+
+- @pnpm/cli-meta@1000.0.4
+
 ## 1000.0.3
 
 ### Patch Changes