+++ fetching/fetcher-base/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @pnpm/fetcher-base
 
+## 1000.0.5
+
+### Patch Changes
+
+- Updated dependencies [a5e4965]
+  - @pnpm/types@1000.2.1
+  - @pnpm/resolver-base@1000.1.4
+
 ## 1000.0.4
 
 ### Patch Changes