+++ fs/find-packages/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pnpm/fs.find-packages
 
+## 1000.0.7
+
+### Patch Changes
+
+- b8b0c68: `fast-glob` replace with `tinyglobby` to reduce the size of the pnpm CLI dependencies [#9169](https://github.com/pnpm/pnpm/pull/9169).
+- Updated dependencies [a5e4965]
+  - @pnpm/types@1000.2.1
+  - @pnpm/read-project-manifest@1000.0.7
+
 ## 1000.0.6
 
 ### Patch Changes