+++ lockfile/filtering/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pnpm/filter-lockfile
 
+## 1001.0.5
+
+### Patch Changes
+
+- Updated dependencies [a5e4965]
+  - @pnpm/types@1000.2.1
+  - @pnpm/dependency-path@1000.0.4
+  - @pnpm/package-is-installable@1000.0.6
+  - @pnpm/lockfile.types@1001.0.4
+  - @pnpm/lockfile.utils@1001.0.4
+  - @pnpm/lockfile.walker@1001.0.4
+
 ## 1001.0.4
 
 ### Patch Changes