+++ lockfile/fs/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @pnpm/lockfile-file
 
+## 1001.1.5
+
+### Patch Changes
+
+- Updated dependencies [a5e4965]
+  - @pnpm/types@1000.2.1
+  - @pnpm/dependency-path@1000.0.4
+  - @pnpm/lockfile.merger@1001.0.4
+  - @pnpm/lockfile.types@1001.0.4
+  - @pnpm/lockfile.utils@1001.0.4
+
 ## 1001.1.4
 
 ### Patch Changes