+++ packages/core-loggers/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pnpm/core-loggers
 
+## 1000.1.4
+
+### Patch Changes
+
+- Updated dependencies [a5e4965]
+  - @pnpm/types@1000.2.1
+
 ## 1000.1.3
 
 ### Patch Changes