+++ pkg-manager/get-context/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @pnpm/get-context
 
+## 1001.0.7
+
+### Patch Changes
+
+- Updated dependencies [a5e4965]
+  - @pnpm/types@1000.2.1
+  - @pnpm/lockfile.fs@1001.1.5
+  - @pnpm/core-loggers@1000.1.4
+  - @pnpm/modules-yaml@1000.1.4
+  - @pnpm/read-projects-context@1000.0.8
+  - @pnpm/resolver-base@1000.1.4
+
 ## 1001.0.6
 
 ### Patch Changes