+++ pnpm/dev/CHANGELOG.md
@@ -1,5 +1,12 @@
 # pd
 
+## 1.0.15
+
+### Patch Changes
+
+- @pnpm/workspace.find-packages@1000.0.13
+- @pnpm/workspace.read-manifest@1000.1.1
+
 ## 1.0.14
 
 ### Patch Changes