+++ reviewing/list/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @pnpm/list
 
+## 1000.0.10
+
+### Patch Changes
+
+- Updated dependencies [a5e4965]
+  - @pnpm/types@1000.2.1
+  - @pnpm/read-package-json@1000.0.5
+  - @pnpm/read-project-manifest@1000.0.7
+  - @pnpm/reviewing.dependencies-hierarchy@1001.0.7
+
 ## 1000.0.9
 
 ### Patch Changes