+++ reviewing/plugin-commands-listing/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @pnpm/plugin-commands-listing
 
+## 1000.0.14
+
+### Patch Changes
+
+- Updated dependencies [a5e4965]
+- Updated dependencies [d965748]
+  - @pnpm/types@1000.2.1
+  - @pnpm/config@1002.5.0
+  - @pnpm/cli-utils@1000.0.13
+  - @pnpm/list@1000.0.10
+
 ## 1000.0.13
 
 ### Patch Changes