+++ store/store-connection-manager/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @pnpm/store-connection-manager
 
+## 1000.0.12
+
+### Patch Changes
+
+- Updated dependencies [d965748]
+  - @pnpm/config@1002.5.0
+  - @pnpm/cli-meta@1000.0.4
+  - @pnpm/client@1000.0.8
+  - @pnpm/package-store@1000.0.7
+  - @pnpm/server@1000.0.7
+
 ## 1000.0.11
 
 ### Patch Changes