+++ store/store-controller-types/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @pnpm/store-controller-types
 
+## 1001.0.3
+
+### Patch Changes
+
+- Updated dependencies [a5e4965]
+  - @pnpm/types@1000.2.1
+  - @pnpm/fetcher-base@1000.0.5
+  - @pnpm/resolver-base@1000.1.4
+
 ## 1001.0.2
 
 ### Patch Changes