+++ workspace/injected-deps-syncer/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @pnpm/workspace.injected-deps-syncer
 
+## 1000.0.1
+
+### Patch Changes
+
+- @pnpm/directory-fetcher@1000.1.1
+- @pnpm/modules-yaml@1000.1.4
+
 ## 1000.0.0
 
 ### Major Changes